pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "uht-dmslibrarian"
version = "0.2.9"
description = "Extension of the UMIC-seq Pipeline - Complete pipeline for dictionary building and NGS count inetgration, with fitness calculations, error modelling and mutation analysis"
readme = "README.md"
requires-python = ">=3.10,<3.11"
license = {text = "MIT"}
authors = [
    {name = "Matt Penner", email = "mp957@cam.ac.uk"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Programming Language :: Python :: 3.10",
]
keywords = ["bioinformatics", "pacbio", "umi", "sequencing", "variant", "calling"]
dependencies = [
    "biopython==1.86",
    "scikit-bio==0.7.0",
    "numpy==2.2.1",
    "pandas==2.3.3",
    "matplotlib==3.10.7",
    "seaborn==0.13.2",
    "scipy==1.15.3",
    "scikit-allel==1.3.13",
    "tqdm==4.67.1",
    "psutil==7.1.3",
    "PyYAML>=6.0",
    "gradio>=4.0.0",
]

[project.scripts]
umic-seq-pacbio = "uht_DMSlibrarian.cli:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["uht_DMSlibrarian*"]

[tool.setuptools.package-data]
uht_DMSlibrarian = ["*.py", "reporting/*.png"]
